feat: 1.适配微服务化的重构 2.个性推荐优化（模型没费用了，不想调了，下次再看）

+++ cloud-common/cloud-common-security/src/main/java/com/cloud/common/security/dto/TokenRefreshRequest.java
@@ -14,18 +14,18 @@ import java.util.Map;
 @Data
 @Builder
 public class TokenRefreshRequest {
-    
+
     private String userKey;
-    
-    private Integer userId;
-    
+
+    private Long userId;
+
     private String username;
-    
+
     private String deviceId;
-    
+
     private DeviceType deviceType;
-    
+
     private Map<String, Object> oldTokenInfo;
-    
+
     private String currentIp;
 }
